Support ppc64le SCV syscall insn
If the kernel supports it AND glibc supports it, the ppc64le
syscalls may use the SCV variant instead of the SC variant.
The registers we care about have the same use so little change
is required.

RHBZ #1922305

This is downstream Fedora patch: ltrace-0.7.91-ppc64le-scv.patch
The downstream patch can be removed.

Signed-off-by: DJ Delorie <dj@redhat.com>