PowerPC ltrace-elf handle LEP/GEP offset
Update logic to increment the address so ltrace will set
it's breakpoint on the Local Entry Point (LEP),
instead of the Global Entry Point (GEP),
so ltrace can properly detect when a function is entered.

This (LEP/GEP) is a feature of the ELF V2 ABI.  The feature was
partially handled with previous patches, but needed a few tweaks.

This patch resolves ~6 errors in testsuite/ltrace.main/branch_func.exp.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>