Avoid infinite loop when single stepping into PLT
When we (ltrace) are single-stepping we need to stop
rather than continue when we determine our current
location matches the libsym entry point.  This change
resolves a condition where ltrace would spin forever
while trying to progress past a breakpoint set on
a lib->entry point.

This fix allows 5 of the ltrace.main/dwarf.exp sub-tests
to run where they previously hung, and results in 5
more test case attempts and 8 more test case passes.

Credit:
This patch was previously submitted to ltrace-devel by
Thierry Fauck, but had not made it into the ltrace
source trees.
  From: Thierry Fauck <tfa...@free.fr>
  Date:   Wed, 23 Mar 2016 11:13:03 +0100
  Subject: [PATCH 1/3] When stepping and plt is UNRESOLVED we
           need to stop to avoid undefinite loop
    This patch return appropriate flag to stop stepping
        modified:   sysdeps/linux-gnu/ppc/plt.c

Signed-off-by:  Will Schmidt <will_schmidt@vnet.ibm.com>