Add -Wno-psabi to avoid compilation warning/failure
The parameters-hfa.exp test uses floating point related
structures as parameters.  The passing of floating point
values changed in GCC 5 for PowerPC, and compilation of
this test code triggers a "note:"
  "note: the ABI of passing homogeneous float
                 aggregates has changed in GCC 5"

This note is enough to cause the compilation to fail
within the test harness, so this test would never run.

Per inspection this warning (along with others) can be
suppressed when the -Wno-psabi option is specified.

There is a pragma that *should* be able to disable the note,
adding this pragma is not sufficient to suppress the note.
  #pragma GCC diagnostic ignored "-Wno-psabi"

This patch adds the -Wno-psabi flag into the
LtraceCompileObjects procedure, with a target check to
limit the effect to powerpc*-*-linux* targets.

This change allows the parameters-hfa.exp source to compile
and run, increasing the number of expected passes by 36, and
decreasing the number of failures by 37.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>