PowerPC Fix HFA FP/GPR parameter logic typo
This corrects a ==/<= typo on an if-condition
in allocate_hfa.  Properly entering this section
of code in turn resolves several testcase failures
in ltrace.main/parameters-hfa.exp where the reported
function parameters were incorrectly displayed.

This patch is based on a patch initially posted by
Thierry Fauck on the ltrace-devel mailing list on 23 Mar 2016.
  From: Thierry Fauck <tfa...@free.fr>
  Date:   Wed, 23 Mar 2016 11:13:05 +0100
  [Typo in ppc64 hfa handling preventing parameters-hfa test
  to pass (bad float register count)

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>